general optimizations

+++ src/index.php
@@ -6,10 +6,13 @@ use skrtdev\Telegram\Message;
 
 require_once("config.php");
 
-if (isset($modules))
+if (isset($modules)) {
     foreach ($modules as $module => $active) {
-        if ($active) require_once($module);
+        if ($active) {
+            require_once($module);
+        }
     }
+}
 
 if (isset($dbname, $dbpass, $dbuser)) {
     try {
@@ -40,7 +43,10 @@ if (isset($Bot)) {
         $start = hrtime(true);
         $lagMex = $message->chat->sendMessage("lag...");
         $time_elapsed_secs = hrtime(true) - $start;
-        $lagMex->editText("elapsed time: " . round($time_elapsed_secs / 1000000, 2) . " ms", true);
+
+        if (!is_null($lagMex)) {
+            $lagMex->editText("elapsed time: " . round($time_elapsed_secs / 1000000, 2) . " ms", true);
+        }
     });
 
     $Bot->onCommand('info', function (Message $message) use ($Bot, $PDO) {
@@ -63,15 +69,15 @@ if (isset($Bot)) {
             try {
                 getUserInfo($Bot, $chatid, $user["user_id"], $user["username"], $user["firstname"], $user["lastname"]);
             } catch (BadRequestException $e) {
-                if (str_contains($e, "invalid user_id specified")) $message->reply("user not present in database");
+                if (str_contains($e, "invalid user_id specified")) {
+                    $message->reply("user not present in database");
+                }
             }
 
+        } else if (is_null($replyToMessage)) {
+            getUserInfo($Bot, $chatid, $userid, $username, $firstName, $lastName);
         } else {
-            if (is_null($replyToMessage)) {
-                getUserInfo($Bot, $chatid, $userid, $username, $firstName, $lastName);
-            } else {
-                getUserInfo($Bot, $chatid, $useridReply, $usernameReply, $firstNameReply, $lastNameReply);
-            }
+            getUserInfo($Bot, $chatid, $useridReply, $usernameReply, $firstNameReply, $lastNameReply);
         }
 
     });
@@ -84,30 +90,34 @@ if (isset($Bot)) {
     });
 
     $Bot->onMessage(function (Message $message) {
-        if (!is_null($message->text) and str_contains($message->text, ":sh")) {
-            // only if message sender is the bot owner
-            if (isBotOwner($message?->from?->id)) {
-                // START: remove :sh and keep every other thing
-                $explodeMessage = explode(" ", $message->text);
-                unset($explodeMessage[0]);
-                $command = implode(" ", $explodeMessage);
-                // END: remove :sh and keep every other thing
+        // only if message sender is the bot owner
+        if (!is_null($message->text) && str_contains($message->text, ":sh") && isBotOwner($message?->from?->id)) {
+            // START: remove :sh and keep every other thing
+            $explodeMessage = explode(" ", $message->text);
+            unset($explodeMessage[0]);
+            $command = implode(" ", $explodeMessage);
+            // END: remove :sh and keep every other thing
 
-                if (str_contains($command, "rm "))
-                    $message->reply("you can only delete files manually");
+            if (str_contains($command, "rm ")) {
+                $message->reply("you can only delete files manually");
+            }
+            else {
+
+                if (PHP_OS_FAMILY === "Windows") {
+                    $response = shell_exec("powershell; " . $command);
+                } // if host os is windows
+                elseif (str_contains($command, "neofetch")) {
+                    $response = shell_exec("neofetch --stdout");
+                } // if host os is !windows and command is neofetch, use --stdout flag else it will show bad output
                 else {
+                    $response = shell_exec($command . " 2<&1");
+                } // any other thing in host os !windows
 
-                    if (PHP_OS_FAMILY === "Windows")
-                        $response = shell_exec("powershell; " . $command); // if host os is windows
-                    elseif (str_contains($command, "neofetch"))
-                        $response = shell_exec("neofetch --stdout"); // if host os is !windows and command is neofetch, use --stdout flag else it will show bad output
-                    else
-                        $response = shell_exec($command . " 2<&1"); // any other thing in host os !windows
-
-                    if (!is_null($response) and str_contains($response, "sh:"))
-                        $message->reply("$command, not found"); // if command is not found
-                    elseif (isset($response))
-                        $message->reply("<pre><b>francesco@Francescos-MBA</b>:<b>~ </b></pre>" . $response);
+                if (!is_null($response) and str_contains($response, "sh:")) {
+                    $message->reply("$command, not found");
+                } // if command is not found
+                elseif (isset($response)) {
+                    $message->reply("<pre><b>francesco@Francescos-MBA</b>:<b>~ </b></pre>" . $response);
                 }
             }
         }
@@ -124,11 +134,12 @@ if (isset($Bot)) {
         $row = $result->fetch(PDO::FETCH_ASSOC);
 
         try {
-            if (isset($row['username']) !== isset($username) and isset($row['firstname']) !== isset($firstName) and isset($row['lastname']) !== isset($lastName)) {
-                $PDO->query("DELETE FROM frasharpbot.user WHERE user.user_id = '$userid'");
-                $PDO->query("INSERT INTO frasharpbot.user (`user_id`, `username`, `firstname`, `lastname`) VALUES ('$userid', '$username', '$firstName', '$lastName')");
-            } elseif (isset($row['id']) != isset($userid))
-                $PDO->query("INSERT INTO frasharpbot.user (`user_id`, `username`, `firstname`, `lastname`) VALUES ('$userid', '$username', '$firstName', '$lastName')");
+            if (isset($row['username']) !== isset($username) && isset($row['firstname']) !== isset($firstName) && isset($row['lastname']) !== isset($lastName)) {
+                $PDO->exec("DELETE FROM frasharpbot.user WHERE user.user_id = '$userid'");
+                $PDO->exec("INSERT INTO frasharpbot.user (`user_id`, `username`, `firstname`, `lastname`) VALUES ('$userid', '$username', '$firstName', '$lastName')");
+            } elseif (isset($row['id']) !== isset($userid)) {
+                $PDO->exec("INSERT INTO frasharpbot.user (`user_id`, `username`, `firstname`, `lastname`) VALUES ('$userid', '$username', '$firstName', '$lastName')");
+            }
         } catch (PDOException) {
         }
     });
@@ -136,7 +147,7 @@ if (isset($Bot)) {
     $Bot->onMessage(function (Message $message) {
         $userid = $message?->from?->id;
         $conn = new mysqli("localhost", getenv("dbuser"), getenv("dbpass"), getenv("dbname"));
-        if ($message->text === ":refresh" and (isAllowed($userid) or isBotOwner($userid))) {
+        if ($message->text === ":refresh" && (isAllowed($userid) || isBotOwner($userid))) {
             $refresh = $message->reply("refreshing...");
 
             $mysqliLogs = $conn->refresh(MYSQLI_REFRESH_LOG);
@@ -151,24 +162,32 @@ if (isset($Bot)) {
 
             $opcache_status = opcache_get_status();
 
-            if (isset($opcache_status["opcache_enabled"]) and !$opcache_status["opcache_enabled"]) {
+            if (isset($opcache_status["opcache_enabled"]) && !$opcache_status["opcache_enabled"]) {
                 $opcacheStatus = "❓ opcache empty";
             } else {
                 $refreshCache = opcache_reset() ? "✅" : "❌";
             }
 
-            $refreshStatus = (
-            ($checkLog === true and $checkTable === true) ? "passed" :
-                (((!$checkLog and $checkTable) or ($checkLog and !$checkTable)) ? "partially passed" : "failed")
-            );
+            if ($checkLog && $checkTable) {
+                $refreshStatus = "passed";
+            } else if (!$checkTable && $checkLog) {
+                $refreshStatus = "partially passed";
+            } else if (!$checkLog && $checkTable) {
+                $refreshStatus = "failed";
+            }
+
             sleep(1);
 
             $refreshText = "refresh status: $refreshStatus";
             $refreshText .= "\n\t$refreshLog database logs";
             $refreshText .= "\n\t$refreshTable database tables";
-            if (isset($refreshCache))
+            if (isset($refreshCache)) {
                 (!isset($opcacheStatus)) ? $refreshText .= "\n\t$refreshCache caches" : $refreshText .= "\n\t$opcacheStatus";
-            $refresh->editText($refreshText);
+            }
+
+            if (!is_null($refresh)) {
+                $refresh->editText($refreshText);
+            }
         }
     });
 
@@ -186,23 +205,21 @@ if (isset($Bot)) {
         if (str_starts_with($message->text, ":kick")) {
             try {
                 $useridToKick = str_ireplace(":kick ", "", $message->text);
-                if ($chat->type == "group") {
+                if ($chat->type === "group") {
                     $message->reply("this can't be used in normal groups");
-                } else {
-                    if ((isBotSudo($userid) or hasRight($userid, $chatid, "can_restrict_members")) and is_null($replyToMessage) and is_null($usernameReply)) {
-                        kickMember($chatid, $useridToKick);
-                        $message->reply($useridToKick . " was kicked");
-                    } else {
-                        if ((isBotSudo($userid) or hasRight($userid, $chatid, "can_restrict_members"))) {
-                            kickMember($chatid, $useridReply);
-                            $message->reply($mentionUserReply . " was kicked");
-                        } else {
-                            if (isAdmin($useridReply, $chatid) and (isBotSudo($userid) or hasRight($userid, $chatid, "can_restrict_members"))) $message->reply("this user is admin, it cannot be kicked");
-                        }
-                    }
+                } else if (is_null($replyToMessage) && is_null($usernameReply) && (isBotSudo($userid) || hasRight($userid, $chatid, "can_restrict_members"))) {
+                    kickMember($chatid, $useridToKick);
+                    $message->reply($useridToKick . " was kicked");
+                } else if ((isBotSudo($userid) || hasRight($userid, $chatid, "can_restrict_members"))) {
+                    kickMember($chatid, $useridReply);
+                    $message->reply($mentionUserReply . " was kicked");
+                } else if (isAdmin($useridReply, $chatid) && (isBotSudo($userid) || hasRight($userid, $chatid, "can_restrict_members"))) {
+                    $message->reply("this user is admin, it cannot be kicked");
                 }
             } catch (BadRequestException $e) {
-                if (str_contains($e, "not enough rights to restrict/unrestrict chat member")) $message->reply("i don't have enough rights");
+                if (str_contains($e, "not enough rights to restrict/unrestrict chat member")) {
+                    $message->reply("i don't have enough rights");
+                }
             }
         }
     });
@@ -221,32 +238,37 @@ if (isset($Bot)) {
                 //$message->reply("you don't have enough rights to warn a user");
             if (isAdmin($useridReply, $chatid)) {
                 $message->reply($mentionUserReply . " is admin, i can't warn them");
-            } elseif (warnMember($chatid, $useridReply, $PDO) and !is_null($useridReply)) {
+            } elseif (warnMember($chatid, $useridReply, $PDO) && !is_null($useridReply)) {
                 $getWarns = $PDO->query("select * from frasharpbot.warns where warns.user_id = '$useridReply' and warns.chat_id = '$chatid'");
                 $currentWarns = $getWarns?->fetch(PDO::FETCH_ASSOC);
-                if ($currentWarns['warns'] <= getMaxWarns($chatid, $PDO))
+                if ($currentWarns['warns'] <= getMaxWarns($chatid, $PDO)) {
                     $message->reply("$mentionUserReply is warned: {$currentWarns['warns']}/" . getMaxWarns($chatid, $PDO));
-                if (!is_null(getMaxWarns($chatid, $PDO)) and $currentWarns['warns'] == getMaxWarns($chatid, $PDO)) {
+                }
+                if (!is_null(getMaxWarns($chatid, $PDO)) && $currentWarns['warns'] === getMaxWarns($chatid, $PDO)) {
                     kickMember($chatid, $useridReply);
                     $message->chat->sendMessage($mentionUserReply . " kicked. " . getMaxWarns($chatid, $PDO) . " warns limit exceeded");
-                    $PDO->query("update frasharpbot.warns set warns.warns = 0 where warns.user_id = '$useridReply' and warns.chat_id = '$chatid'");
+                    $PDO->exec("update frasharpbot.warns set warns.warns = 0 where warns.user_id = '$useridReply' and warns.chat_id = '$chatid'");
                 }
             }
 
         } catch (BadRequestException $e) {
-            if (str_contains($e, "not enough rights to restrict/unrestrict chat member")) $message->reply("i don't have enough rights");
+            if (str_contains($e, "not enough rights to restrict/unrestrict chat member")) {
+                $message->reply("i don't have enough rights");
+            }
         }
     });
 
     $Bot->onMessage(function (Message $message) use ($PDO) {
         if (str_starts_with($message->text, ":setMaxWarns")) {
             $getWarnsVal = explode(" ", $message->text);
-            $maxWarns = intval($getWarnsVal[1]);
+            $maxWarns = (int)$getWarnsVal[1];
 
-            if ($maxWarns > 0 and $maxWarns <= 10)
+            if ($maxWarns > 0 && $maxWarns <= 10) {
                 setMaxWarns($message->chat->id, $maxWarns, $PDO);
+            }
         }
     });
+
 
     $Bot->addErrorHandler(function ($e) {
         print('Caught ' . get_class($e) . ' exception from general handler' . PHP_EOL);
